refactor(mentor): entity내 패키지 종류별로 카테고리화
