+++ src/main/java/com/sprint/mission/discodeit/entity/base/BaseEntity.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.entity;
+package com.sprint.mission.discodeit.entity.base;
 
 import java.util.UUID;
