+++ src/main/java/com/sprint/mission/discodeit/entity/message/ChannelMessage.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.entity;
+package com.sprint.mission.discodeit.entity.message;
 
 import java.util.UUID;
