+++ src/main/java/com/sprint/mission/discodeit/service/UserService.java
@@ -1,6 +1,6 @@
 package com.sprint.mission.discodeit.service;
 
-import com.sprint.mission.discodeit.entity.User;
+import com.sprint.mission.discodeit.entity.user.User;
 
 import java.util.List;
 import java.util.UUID;
